backend/requirements.txt
fastapi==0.99.1
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
selenium==4.21.0
openai==1.30.1
google-generativeai==0.7.0
msgraph-sdk==1.3.0
python-dotenv==1.0.1
pydantic<2.0
python-multipart==0.0.9
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
psycopg2-binary==2.9.9
email-validator
aiohttp==3.9.1
psutil==5.9.6
